==> Progetto/02 Data Cleaning - Categorical Data Transformation.ipynb <==
<center>

# Web Intelligence

# Who will win the Australian Open 2020 Tournament?

## 03 Data Cleaning - Categorical Data Transformation

### Riccardo Spolaor (864877)

</center>

# Trasformazione Variabili Categoriali
Controllo solo le feature categoriali (Tipo = *Object*) e procedo alla loro trasformazione in variabili numeriche.

import pandas as pd
from sklearn.preprocessing import LabelEncoder

dataframe = pd.read_csv('./datasets/dataframe01.csv', low_memory = False)
dataframe.head()

print ("Variabili Categoriali:")
print(dataframe.dtypes[dataframe.dtypes == object].index.values)

for value in dataframe.dtypes[dataframe.dtypes == object].index.values:
    print('\n', value, 'values length: ', len(dataframe[value].unique()))
    print('\t', dataframe[value].unique()[:20])

## One-Hot-Encoding
Uso la tecnica dello ***one-hot-encoding*** per trasformare le feature con un numero limitato di label e che non presentano nessun tipo di struttura gerarchica (***Surface***, ***Comment***). 

Uso il ***Label Encoder*** per codificare ***Court*** a 0 o 1, dato che contiene solo variabili.

leCourt = LabelEncoder().fit(dataframe['Court'])

dataframe['Court'] = leCourt.transform(dataframe['Court'])


dataframe = pd.get_dummies(dataframe, prefix = ['', ''],
                   prefix_sep = '', columns = ['Surface', 'Comment'])

dataframe.columns

for col in ['Clay', 'Grass', 'Hard', 'Awarded', 'Completed', 'Disqualified','Retired', 'Sched', 'Walkover']:
    print (col,'number of entries:', dataframe[dataframe[col]==1].shape[0])

## Feature che presentano un Ordine

### Round
Per quanto riguarda i ***Round***:
* Posso assegnare alle feature un valore numerico crescente in base all'importanza del Round:
    * *1st Round* < *2nd Round* < *3rd Round* < *4th Round* < *Round Robin* < *Quarterfinals* < *Semifinals* < *The Final*;

print('Tipi di Round:')

print(dataframe['Round'].unique())

roundsOrdered = {'1st Round': 0, '2nd Round': 1, 'Quarterfinals': 5, 'The Final': 7, 'Semifinals': 6,
 '3rd Round': 2, '4th Round': 3, 'Round Robin': 4}

dataframe['Round'] = dataframe['Round'].apply(lambda x : roundsOrdered[x])

### Series

Per quanto riguarda la colonna ***Series***:

* Posso assegnare alle feature un valore numerico crescente in base all'importanza del tipo di torneo:
    * *ATP 250 = International* < *ATP 500 = International Gold* < *Masters 1000 = Masters* < *Masters Cup = ATP Finals* < *Grand Slam*

print('Tipi di Tornei:')

print(dataframe['Series'].unique())

pd.DataFrame(columns = ['Series','WPts'], data = [
    ['Grand Slam', 2000],
    ['Masters Cup/ATP Finals', 1500],
    ['ATP250/International', 250],
    ['ATP500/International Gold', 500],
    ['Masters/Masters 1000', 1000],
])

seriesOrdered = {'ATP250': 0, 'International': 0, 'ATP500': 1,'International Gold': 1, 'Masters' : 2, 'Masters 1000': 2,
                'Masters Cup': 3, 'Grand Slam': 4}
dataframe['Series'] = dataframe['Series'].apply(lambda x : seriesOrdered[x])

### Tournament
La feature ***Tournament*** presenta troppi label distinti.
Per non allargare troppo il Dataframe con lo ***one-hot-encoding*** su tutti i label:
* Seleziono i primi 25 Tornei più prestigiosi e li categorizzo in ordine crescente partendo da 1;
* Gli altri li categorizzo con 0.

print('Grand Slam: ', dataframe[dataframe['Series']==4]['Tournament'].unique(), '\n')
print('Masters Cup: ', dataframe[dataframe['Series']==3]['Tournament'].unique(), '\n')
print('Masters 1000: ', dataframe[dataframe['Series']==2]['Tournament'].unique()[:20], '\n')
print('ATP500: ', dataframe[dataframe['Series']==1]['Tournament'].unique()[:20], '\n')
print('ATP250: ', dataframe[dataframe['Series']==0]['Tournament'].unique()[:20], '\n')

Includo tutti i tipi di *Grand Slam*, il torneo *Masters Cup*, tutti i tornei del tipo *Masters 1000* e i primi 10 tornei dell'*ATP500* classificati secondo **Tennis365** e **Forbes**

Fonti: <br>
* https://www.tennis365.com/tennis-top-10/top-10-atp-tour-500-series-events-where-does-queens-club-rank/
* https://www.tennis365.com/tennis-top-10/top-10-biggest-non-grand-slam-tournaments-on-the-tennis-tour/
* https://www.forbes.com/sites/monteburke/2012/05/30/what-is-the-most-prestigious-grand-slam-tennis-tournament/#2370e5097696

def encodeTournaments():
    grandSlams = ['Wimbledon', 'US Open', 'French Open', 'Australian Open']
    mastersCup = ['Masters Cup']
    masters1000 = ['Rogers Masters','BNP Paribas Masters', 'Shanghai Masters', 'Monte Carlo Masters', 
                   'Western & Southern Financial Group Masters', 
                   'Sony Ericsson Open', 'Mutua Madrid Open', 'Mutua Madrileña Madrid Open', 
                   'Internazionali BNL d\'Italia', 'BNP Paribas Open']
    atp500best = ['Citi Open','Halle Open', 'Dubai Tennis Championships', 'China Open', 'ABN AMRO World Tennis Tournament', 
              'Swiss Indoors', 'Rakuten Japan Open Tennis Championships','German Open Tennis Championships',
              'Open Banco Sabadell', 'Queen\'s Club Championships']
    
    tournamentsDict = {t:i+1 for i,t in enumerate(atp500best[::-1] + masters1000[::-1] + mastersCup +
                                                  grandSlams[::-1])}
    
    
    dataframe['Tournament'] = dataframe['Tournament'].apply(lambda x : 
                                                               tournamentsDict[x] if x in tournamentsDict
                                                               else 0)
    return tournamentsDict

tournamentsDict = encodeTournaments()

### Location
Da quanto notato nell'analisi del dataset, possiamo fare l'encoding delle location considerndo il loro ordine in base al numero di match giocati.
Utilizzare lo ***one-hot-encoding*** allargherebbe troppo il dataset.

ordered_locations = {l:i for i, l in enumerate(dataframe['Location'].value_counts(ascending = True).index)}
dataframe['Location'] = dataframe['Location'].apply(lambda x: ordered_locations[x])

### Winner e Loser

Trasformare le feature ***Winner*** e ***Loser*** con lo **one-hot-encoding** allargherebbe troppo il dataframe:
* Ordino i giocatori in base alla differenza tra il numero delle loro vittorie e quello delle loro sconfitte;
* Creo delle colonne per i migliori 25 giocatori per il giocatore A e il giocatore B che assumono i valori:
    * 1 se il giocatore è il giocatore indicato nella colonna;
    * 0 altrimenti
* Aggiungo una colonna *OtherA* e *OtherB* per indicare se il giocatore non è tra i top 25.

def setBest25Players():
    players = dataframe['Winner'].value_counts().subtract(
        dataframe['Loser'].value_counts(), fill_value = 0).sort_values(ascending = False).index.tolist()[:25]

    for p in players:
        dataframe[p+'A'] = dataframe['Winner'].apply(lambda x: int(x == p))
        dataframe[p+'B'] = dataframe['Loser'].apply(lambda x: int(x == p))
        
    dataframe['OtherA'] = dataframe['Winner'].apply(lambda x: int(x not in players[:25]))
    dataframe['OtherB'] = dataframe['Loser'].apply(lambda x: int(x not in players[:25]))
    
    return players

best_25_players = setBest25Players()

dataframe.head()

%store best_25_players
%store ordered_locations
%store tournamentsDict

dataframe.to_csv('./datasets/dataframe02.csv', index = False)
