==> Progetto/05 Adding New Features.ipynb <==
<center>

# Web Intelligence

# Who will win the Australian Open 2020 Tournament?

## 02 Adding New Features

### Riccardo Spolaor (864877)

</center>

# Creazione Variabili aggiuntive
Viene ripreso il dataset usato prima del Notebook *03 Building The Prediction Model* per poter aggiungere nuove features basate sui dati non utilizzabili nel modello di predizione.

import pandas as pd
import python_files.FeaturesOperations as featop

dataframe = pd.read_csv('./datasets/dataframe02.csv', low_memory = False)
dataframe = dataframe.reset_index(drop = True)

dataframe.head()

## Feature di Confronto
Vengono aggiunte feature di confronto di maggioranza tra le caratteristiche dei due giocatori (es: ***Rank***; ***Pts***).

Le feature di confronto verranno aggiunte anche per le nuove variabili aggiunte.

dataframe['RankA>RankB'] = dataframe.apply(lambda x : int(x['WRank'] > x['LRank']), axis = 1)
dataframe['MaxA>MaxB'] = dataframe.apply(lambda x : int(x['MaxW'] > x['MaxL']), axis = 1)
dataframe['AvgA>AvgB'] = dataframe.apply(lambda x : int(x['AvgW'] > x['AvgL']), axis = 1)
dataframe['PtsA>PtsB'] = dataframe.apply(lambda x : int(x['WPts'] > x['LPts']), axis = 1)

## Caratteristiche dello storico dei due giocatori
* ***OpponentsPlayed***: Il numero di partite giocate tra i due contendenti;
* ***OpponentsWRatio(A/B)***: Il rapporto vittorie/partite giocate tra i due contendenti per i due giocatori;

featop.set_player_a_vs_player_b_statistics(dataframe)

(dataframe[dataframe['Winner'] == 'NADAL R.'].merge(
    dataframe[dataframe['Loser'] == 'FEDERER R.'])).append(
    dataframe[dataframe['Winner'] == 'FEDERER R.'].merge(
    dataframe[dataframe['Loser'] == 'NADAL R.'])).sort_values(by = 'OpponentsPlayed')[['Winner', 'Loser', 'OpponentsPlayed',
                                                                                       'OpponentsWRatioA', 'OpponentsWRatioB', 
                                                                                       'OpponentsWRatioA>OpponentsWRatioB', 
                                                                                       'Date']]

## Feature riguardanti le caratteristiche dei giocatori nel torneo
* ***FadigueTournGames(A/B)***: i game giocati finora da un giocatore nel torneo come misura di stanchezza;
* ***FadigueTournSets(A/B)***: i set giocati finora da un giocatore nel torneo come misura di stanchezza;

featop.set_fatigue(dataframe)

dataframe[dataframe['csvID'] == 3].merge(
    dataframe[dataframe['ATP'] == 5]).merge(
    dataframe[dataframe['Winner'] == 'HEWITT L.']).head(20)[['Winner','Loser', 'Round','W1', 'W2', 'W3', 'Wsets',
                                               'L1', 'L2', 'L3', 'Lsets','FadigueTournGamesA','FadigueTournGamesB',
                                               'FadigueTournGamesA>FadigueTournGamesB', 'FadigueTournSetsA','FadigueTournSetsB',
                                               'FadigueTournSetsA>FadigueTournSetsB']]

## Feature riguardanti le caratteristiche del giocatore in generale
* ***RetiredLast(A/B)***: feature binaria che indica se un giocatore si è ritirato lo scorso match;
* ***WalkoverLast(A/B)***: feature binaria che indica se un giocatore si è infortunato lo scorso match;
* ***Played(A/B)***: partite giocate dal giocatore finora;
* ***WonRatio(A/B)***: rapporto partite vinte/giocate dal giocatore finora.
* ***PlayedCourt(A/B)***: partite giocate dal giocatore nel tipo di campo del match attuale finora;
* ***WonRatioCourt(A/B)***: rapporto partite vinte/giocate dal giocatore nel tipo di campo del match attuale finora.
* ***5_gamesMean(A/B)***: media del numero di game vinti dal giocatore negli ultimi 5 match;
* ***5_setsMean(A/B)***: media del numero di set vinti dal giocatore negli ultimi 5 match;

featop.set_retired_walkover_last(dataframe)

first_nadal_walkover = dataframe[dataframe['Walkover'] == 1].merge(dataframe[dataframe['Loser'] == 'NADAL R.']).head(1)['Date']

dataframe[dataframe['Winner'] == 'NADAL R.'].append(
    dataframe[dataframe['Loser'] == 'NADAL R.']).merge(
    dataframe[dataframe['Date'] >= first_nadal_walkover[0]]).sort_values(by = 'Date')[['Winner','Loser','WalkoverLastA', 
                                                                                'WalkoverLastB', 'Walkover', 'Date']].head(5)

featop.set_played_matches(dataframe)

dataframe[dataframe['Winner'] == 'NADAL R.'].append(
    dataframe[dataframe['Loser'] == 'NADAL R.']).sort_values(by='Date').head(20)[['Winner', 'Loser', 'PlayedA', 'PlayedB', 
                                                                                  'Date','WonRatioA','WonRatioB']]

print(dataframe[dataframe['Winner'] == 'NADAL R.'].append(dataframe[dataframe['Loser'] == 'NADAL R.']).shape[0])

dataframe[dataframe['Winner'] == 'NADAL R.'].append(
    dataframe[dataframe['Loser'] == 'NADAL R.']).sort_values(by='Date').tail(1)[['Winner', 'Loser', 'PlayedA', 'PlayedB', 'Date',
                                                                               'WonRatioA','WonRatioB']]

featop.set_played_matches_by_court(dataframe)

dataframe[dataframe['Winner'] == 'NADAL R.'].append(
    dataframe[dataframe['Loser'] == 'NADAL R.']).merge(
    dataframe[dataframe['Hard'] == 1]).sort_values(by='Date').head(10)[['Winner', 'Loser', 'PlayedCourtA', 'PlayedCourtB', 'Date',
                                                                               'WonRatioCourtA','WonRatioCourtB']]

print(dataframe[dataframe['Winner'] == 'NADAL R.'].append(dataframe[dataframe['Loser'] == 'NADAL R.']).merge(
    dataframe[dataframe['Hard'] == 1]).shape[0])

dataframe[dataframe['Winner'] == 'NADAL R.'].append(
    dataframe[dataframe['Loser'] == 'NADAL R.']).merge(
    dataframe[dataframe['Hard'] == 1]).sort_values(
        by='Date').tail(1)[['Winner', 'Loser', 'PlayedCourtA', 'PlayedCourtB', 'Date', 'WonRatioCourtA','WonRatioCourtB']]

featop.set_players_last_5_statistics(dataframe)

display(dataframe[dataframe['Winner'] == 'NADAL R.'].append(
            dataframe[dataframe['Loser'] == 'NADAL R.']).sort_values(by='Date').head(10)[['Winner', 'Loser',
                                                                                  'W1', 'W2', 'W3', 'W4', 'W5', 'Wsets',
                                                                                  '5_gamesMeanA', '5_setsMeanA',
                                                                                  'L1', 'L2', 'L3', 'L4', 'L5', 'Lsets',
                                                                                  '5_gamesMeanB', '5_setsMeanB',
                                                                                  '5_gamesMeanA>5_gamesMeanB',
                                                                                  '5_setsMeanA>5_setsMeanB']])

# Data Integration
Aggiungo la mano con cui gioca un giocatore sfruttando un dataset esterno.
Aggiungo poi le features riguardanti le partite giocate contro i giocatori che hanno la stessa mano dell'avversario e il rapporto partite vinte/partite giocate contro i giocatori che hanno la stessa mano dell'avversario.
Nuove feature:
* ***Hand(A/B)***: mano con cui gioca il giocatore;
* ***PlayedVsSameHanded(A/B)***: Il numero di partite disputate dal giocatore A o B contro giocatori che usano la stessa mano dell’avversario prima del match;
* ***WonRatioVsSameHanded(A/B)***: Il rapporto tra partite vinte dal giocatore A o B e giocate contro giocatori che usano la stessa mano dell’avversario prima del match; 


Dataset preso da:
* https://www.kaggle.com/sijovm/atpdata

Valori mancanti recuperati da:
* https://www.ultimatetennisstatistics.com/
* https://www.tennisexplorer.com/

featop.set_players_hands(dataframe)

print('Valori di HandA:',dataframe['HandA'].unique())
print('Valori di HandB:',dataframe['HandB'].unique())

featop.set_hands_statistics(dataframe)

dataframe[dataframe['Winner'] == 'PUERTA M.'].append(
            dataframe[dataframe['Loser'] == 'PUERTA M.']).sort_values(by='Date').head(20)[['Winner', 'Loser',
                                                                                  'HandA', 'HandB', 'PlayedVsSameHandedA',
                                                                                         'PlayedVsSameHandedB',
                                                                                         'WonRatioVsSameHandedA',
                                                                                         'WonRatioVsSameHandedB',
                                                                                         'WonRatioVsSameHandedA>WonRatioVsSameHandedB']]

dataframe.to_csv('./datasets/dataframe05.csv', index = False)
