==> Progetto/06 ReBuilding The Prediction Dataset.ipynb <==
<center>

# Web Intelligence

# Who will win the Australian Open 2020 Tournament?

## 05 Rebuilding The Prediction Dataset

### Riccardo Spolaor (864877)

</center>

Modifico il dataframe in modo da poter ottenere un modello adatto alla predizione:
* Rinomino ***Winner*** in ***PlayerA*** e ***Loser*** in ***PlayerB***;
* Rinomino tutte le colonne con informazioni legate al vincitore e al perdente in modo da riflettere il legame ai nuovi nomi assegnati;
* Creo la colonna ***Winner*** che assume i valori:
    * 0 se ha vinto ***PlayerA***;
    * 1 se ha vinto ***PlayerB***;
* Assegno a ***Winner*** tutti valori 0;
* Duplico il dataframe e faccio lo swapping tra tutte le colonne con informazioni sul Giocatore A e quelle con informazioni sul Giocatore B;
* Assegno il valore 1 a ogni entry della colonna ***Winner*** del nuovo dataframe;
* Unisco i due dataframe.

import pandas as pd
import python_files.ModelBuilder as modbuild

dataframe = pd.read_csv('./datasets/dataframe05.csv', low_memory = False)
%store -r best_25_players

dataframe.columns.values

modbuild.rename_columns(dataframe)

exdataframe = dataframe.copy()

dataframe = modbuild.get_inverted_dataFrame(dataframe,
                                   ['Player', 'Rank', 'RankFilled', 'Pts', 'PtsFilled', '1', '2', '3', '4', '5', 'sets', 
                                    'Max', 'Avg', 'MaxFilled', 'AvgFilled',  'Played', 'WonRatio','WonRatioCourt', 
                                    'PlayedCourt', 'OpponentsWRatio', 'FadigueTournGames', 'FadigueTournSets', 
                                    'WalkoverLast', 'RetiredLast','5_gamesMean', '5_setsMean', 'Hand', 'PlayedVsSameHanded',
                                    'WonRatioVsSameHanded','Other'] + best_25_players)

modbuild.recalculate_comparisons(dataframe)

with pd.option_context('display.max_columns',None):
    display(dataframe.tail())
    display(exdataframe.tail())

dataframe.drop(['Date', 'PlayerA', 'PlayerB', '1A', '1B', '2A', '2B', '3A', '3B', '4A', 
                '4B', '5A', '5B', 'setsA', 'setsB', 'Awarded', 'Completed',
       'Disqualified', 'Retired', 'Sched', 'Walkover'], axis = 1, inplace = True)

dataframe.to_csv('./datasets/dataframe06.csv', index = False)
